# apps/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Salud API
  version: 0.1.0
  description: |
    API de la plataforma de salud digital. Cubre el CRUD de profesionales, su
    disponibilidad, y la autenticación por sesión. Sin persistencia real:
    todo vive en memoria y se pierde al reiniciar.

    La sesión viaja en una cookie `HttpOnly` llamada `sesion`. Las operaciones
    marcadas con `security` la exigen; el resto es público y se lee sin estar
    logueado. Un 401 significa "no sé quién sos"; un 403, "sé quién sos y no
    es tuyo".

    Todo cuerpo tiene que llegar con `Content-Type: application/json`. No es
    formalidad: junto con `SameSite=Lax` es lo que impide que un formulario de
    otro sitio forje una escritura con la cookie del usuario.

    Los montos viajan como enteros en centavos. Nunca como decimales: un
    número de JSON se convierte en float64 del otro lado.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
paths:
...
  /api/v1/usuarios:
    post:
      tags: [autenticacion]
      summary: Registra un usuario
      security: []   # pública
      description: |
        Registrarse deja la sesión abierta: la respuesta trae la cookie. Pedir
        las credenciales otra vez inmediatamente después no informa nada.

        El usuario no es el perfil profesional. Para atender hay que crear
        además un perfil con `POST /api/v1/profesionales`; para reservar
        turnos no hace falta.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionRegistro' }
      responses:
        '201':
          description: Creado, con la sesión abierta
          headers:
            Set-Cookie:
              description: Cookie de sesión `HttpOnly`, `SameSite=Lax`, 7 días
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '409': { $ref: '#/components/responses/EmailEnUso' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
  /api/v1/usuarios/yo:
    get:
      tags: [autenticacion]
      summary: Devuelve el usuario de la sesión
      description: |
        `perfilProfesionalId` es lo único que el cliente necesita para saber
        si mostrar la vista de profesional o la de paciente. Es derivado: no
        hay ningún campo de rol, justamente para que no pueda desincronizarse
        del perfil real.
      security:
        - sesionCookie: []
      responses:
        '200':
          description: El usuario autenticado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioActual' }
        '401': { $ref: '#/components/responses/NoAutenticado' }

    put:
      tags: [autenticacion]
      summary: Cambia los datos personales de la propia cuenta
      description: |
        No recibe un id: el usuario sale de la sesion, asi que no hay forma de
        editar la cuenta de otra persona.

        Sin contrasena. Cambiarla es otra operacion, con otra regla -hay que
        pedir la actual- y aceptarla aca haria que un formulario de datos
        personales pueda reemplazar una credencial.

        Cambiar el email cambia con que se entra, y las sesiones abiertas
        siguen valiendo: el token identifica por id, no por email.
      security:
        - sesionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionPerfil' }
      responses:
        '200':
          description: Datos actualizados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '409':
          description: Ya hay otra cuenta con ese email
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problema' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
  /api/v1/sesiones:
    post:
      tags: [autenticacion]
      summary: Inicia sesión
      security: []   # pública
      description: |
        El 401 es idéntico para "ese email no existe" y "esa contraseña está
        mal", a propósito: distinguirlos convertiría a este endpoint en un
        oráculo de qué direcciones están registradas.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionLogin' }
      responses:
        '201':
          description: Sesión abierta
          headers:
            Set-Cookie:
              description: Cookie de sesión `HttpOnly`, `SameSite=Lax`, 7 días
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '401': { $ref: '#/components/responses/CredencialesInvalidas' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }
...
  /api/v1/turnos:
    get:
      tags: [turnos]
      summary: Mis turnos como paciente
      description: |
        Devuelve los turnos del usuario de la sesión, cancelados incluidos.

        **No acepta un `pacienteId` como filtro**, a propósito: el paciente es
        siempre el de la sesión. Aceptarlo convertiría este endpoint en una
        forma de leer la agenda de cualquiera.
      security:
        - sesionCookie: []
      parameters:
        - name: desde
          in: query
          schema: { type: string, format: date }
          description: |
            Opcional, AAAA-MM-DD, igual que en `/bloqueos`. Sin rango se
            devuelven los vigentes y futuros.
        - name: hasta
          in: query
          schema: { type: string, format: date }
          description: Opcional, AAAA-MM-DD. El día pedido entra entero.
      responses:
        '200':
          description: Los turnos del paciente, con el profesional de cada uno
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaTurnosDePaciente' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
  /api/v1/turnos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    delete:
      tags: [turnos]
      summary: Cancela un turno
      description: |
        Lo pueden hacer **las dos partes**: el paciente y el profesional. Por eso
        la ruta no cuelga de `/profesionales/{id}` — no es una operación del
        dueño de la agenda.

        No es idempotente: cancelar un turno ya cancelado da 422. A diferencia
        del logout, repetir la operación probablemente significa que el cliente
        cree algo distinto de lo que pasó.

        Un turno que ya empezó tampoco se cancela: reescribirlo hacia atrás
        borraría el registro de una consulta que ocurrió.
      security:
        - sesionCookie: []
      responses:
        '204':
          description: Turno cancelado
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
  /api/v1/profesionales:
    get:
      tags: [profesionales]
      summary: Lista profesionales
      security: []   # pública
      description: |
        Por defecto devuelve solo los activos. Para ver los dados de baja hay
        que pedir `estado=inactivo` explícitamente.
      parameters:
        - name: especialidad
          in: query
          schema: { $ref: '#/components/schemas/Especialidad' }
        - name: zona
          in: query
          description: Compara sin distinguir mayúsculas ni acentos
          schema: { type: string }
        - name: estado
          in: query
          schema: { $ref: '#/components/schemas/Estado' }
        - name: busqueda
          in: query
          description: |
            Busca en nombre y apellido, sin distinguir mayúsculas ni acentos.
            Buscar `gonzalez` encuentra a `González`.
          schema: { type: string }
        - name: limite
          in: query
          description: |
            Un valor por encima del máximo no se rechaza: se recorta al
            máximo, y `paginacion.limite` en la respuesta informa el que de
            verdad se aplicó. Paginar sumando el valor pedido en vez del
            devuelto saltea registros.
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: desplazamiento
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaProfesionales' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }

    post:
      tags: [profesionales]
      summary: Da de alta un profesional
      security:
        - sesionCookie: []
      description: |
        El profesional nace con `estado: activo` y `verificacion: pendiente`.
        Nadie nace verificado: la verificación es un acto contra REFEPS que
        todavía no está implementado.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionProfesional' }
      responses:
        '201':
          description: Creado
          headers:
            Location:
              description: URL del recurso creado
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profesional' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '409': { $ref: '#/components/responses/Conflicto' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }
  /api/v1/profesionales/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [profesionales]
      summary: Trae un profesional por ID
      security: []   # pública
      description: |
        Un profesional dado de baja devuelve 200 con `estado: inactivo`. El
        recurso existe, simplemente no opera.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profesional' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

    put:
      tags: [profesionales]
      summary: Reemplaza los campos editables
      security:
        - sesionCookie: []
      description: |
        Reemplazo total, no parcial. `id`, `slug`, `estado`, `verificacion` y
        las marcas de tiempo no son editables.

        El slug no se regenera al cambiar el nombre: es una URL pública y
        romperla rompe enlaces y posicionamiento.

        Cambiar la matrícula o la especialidad devuelve `verificacion` a
        `pendiente`: la verificación se hizo sobre esos datos.

        Funciona sobre profesionales inactivos y no cambia su estado.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionProfesional' }
      responses:
        '200':
          description: Actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profesional' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }

    delete:
      tags: [profesionales]
      summary: Da de baja al profesional
      security:
        - sesionCookie: []
      description: |
        Baja lógica, no borrado. Pone `estado` en `inactivo` y sella
        `dadoDeBajaEn`. El registro se queda: los turnos y comprobantes
        históricos apuntan a él.

        Es idempotente: dar de baja algo ya dado de baja devuelve 204.
      responses:
        '204': { description: Dado de baja }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
  /api/v1/profesionales/{id}/reactivar:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [profesionales]
      summary: Revierte la baja
      security:
        - sesionCookie: []
      description: Idempotente. Sobre alguien ya activo devuelve 200 sin cambios.
      responses:
        '200':
          description: Reactivado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profesional' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
  /api/v1/perfiles/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [perfiles]
      summary: Trae el perfil público de un profesional por su slug
      security: []   # pública
      description: |
        Es lo que consume la página `/p/{slug}` del frontend. Vive en su
        propio recurso, distinto de `profesionales`, porque es de lectura
        pública y no expone lo administrativo del CRUD.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profesional' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /api/v1/profesionales/{id}/horarios:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [profesionales]
      summary: El horario semanal del profesional
      security: []   # pública
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaHorarios' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

    put:
      tags: [profesionales]
      summary: Reemplaza el horario semanal completo
      security:
        - sesionCookie: []
      description: |
        Reemplazo total: la semana se edita como una unidad. Una lista vacía
        deja al profesional sin horarios, que es legítimo.

        Se rechaza con 422 si dos bloques del mismo día se solapan, si en algún
        bloque no entra ninguna sesión, o si algún bloque usa una modalidad que
        el profesional no declara ofrecer.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionHorarios' }
      responses:
        '200':
          description: La semana resultante, ordenada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaHorarios' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }
  /api/v1/profesionales/{id}/bloqueos:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [profesionales]
      summary: Los bloqueos del profesional
      security: []   # pública
      parameters:
        - name: desde
          in: query
          schema: { type: string, format: date }
        - name: hasta
          in: query
          schema: { type: string, format: date }
      description: |
        Sin `desde` ni `hasta` devuelve los vigentes y futuros. Con rango,
        devuelve los que lo pisan, aunque sea en parte.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaBloqueos' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

    post:
      tags: [profesionales]
      summary: Bloquea un rango de la agenda
      security:
        - sesionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionBloqueo' }
      responses:
        '201':
          description: Creado
          headers:
            Location:
              description: URL del recurso creado
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BloqueoCreado' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }
  /api/v1/profesionales/{id}/bloqueos/{bloqueoId}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: bloqueoId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [profesionales]
      summary: Elimina un bloqueo
      security:
        - sesionCookie: []
      description: |
        El bloqueo tiene que ser de ese profesional. Borrar el de otro desde
        esta ruta es un 404.
      responses:
        '204': { description: Eliminado }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
  /api/v1/profesionales/{id}/huecos:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: desde
        in: query
        required: true
        schema: { type: string, format: date }
      - name: hasta
        in: query
        required: true
        schema: { type: string, format: date }
    get:
      tags: [profesionales]
      summary: Los huecos reservables del profesional
      security: []   # pública
      description: |
        `desde` y `hasta` son fechas y las dos entran: del 25 al 27 incluye los
        tres días.

        `hasta` anterior a `desde` es 400: es un query param mal formado, no
        una entidad semánticamente inválida.

        El rango se recorta al horizonte del profesional, no se rechaza, y
        `rango` en la respuesta informa el que de verdad se usó. También se
        recorta hacia atrás: pedir una fecha pasada nunca produce huecos, así
        que `desde` nunca queda antes de hoy.

        Un profesional dado de baja devuelve una lista vacía, no un error.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaHuecos' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
  /api/v1/profesionales/{id}/turnos:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [turnos]
      summary: La agenda ocupada del profesional
      description: |
        **Privado, a diferencia de `/horarios` y `/bloqueos`.** La asimetría es
        deliberada: los huecos libres son información de oferta y son públicos,
        pero la agenda *ocupada* dice quién es paciente de quién, y eso es dato
        de salud bajo Ley 25.326.

        Solo lo responde el dueño del perfil.
      security:
        - sesionCookie: []
      parameters:
        - name: desde
          in: query
          schema: { type: string, format: date }
          description: |
            Opcional, AAAA-MM-DD, igual que en `/bloqueos`. Sin rango se
            devuelven los vigentes y futuros.
        - name: hasta
          in: query
          schema: { type: string, format: date }
          description: Opcional, AAAA-MM-DD. El día pedido entra entero.
      responses:
        '200':
          description: La agenda del profesional, con el paciente de cada turno
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaTurnosDeProfesional' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

    post:
      tags: [turnos]
      summary: Reserva un turno
      description: |
        El paciente sale de la sesión, nunca del cuerpo.

        El `inicio` tiene que coincidir exactamente con un hueco libre de
        `GET .../huecos`. Si dos pacientes piden el mismo al mismo tiempo, el
        segundo recibe 409.
      security:
        - sesionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeticionTurno' }
      responses:
        '201':
          description: Turno reservado
          headers:
            Location:
              description: URL del turno creado
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Turno' }
        '400': { $ref: '#/components/responses/PeticionInvalida' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/TurnoNoDisponible' }
        '413': { $ref: '#/components/responses/CuerpoDemasiadoGrande' }
        '415': { $ref: '#/components/responses/TipoDeContenidoInvalido' }
        '422': { $ref: '#/components/responses/ValidacionFallida' }
components:
  securitySchemes:
    sesionCookie:
      type: apiKey
      in: cookie
      name: sesion
      description: |
        Cookie opaca de sesión, `HttpOnly` y `SameSite=Lax`, con 7 días de
        vigencia absoluta. La setea el registro o el login y el browser la
        manda sola. No es un JWT: el servidor guarda el hash del token, así
        que cerrar sesión la invalida de verdad.
  schemas:
    PeticionRegistro:
      type: object
      required: [email, contrasena, nombre, apellido, telefono]
      properties:
        email: { type: string, format: email, maxLength: 254, example: juan@ejemplo.com }
        contrasena:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: |
            Mínimo 8 caracteres, el piso de NIST SP 800-63B. El máximo son 72
            **bytes**, no caracteres: bcrypt descarta en silencio todo lo que
            pase de ahí, y con acentos el byte 72 llega antes.

            Sin reglas de composición a propósito: alargan el formulario sin
            agregar entropía real.
        telefono:
          type: string
          example: 11 1234-5678
          description: |
            Acepta las formas que se usan de verdad y las normaliza a E.164:
            "11 1234-5678", "011 15 1234-5678" y "+54 9 11 1234 5678" son el
            mismo numero y quedan guardados igual.

            El 0 de larga distancia y el 15 de movil no son parte del numero:
            se sacan. Diez digitos pelados, sin ninguno de esos prefijos, se
            toman como movil — el campo pide un celular, y quien quiera dar un
            fijo lo dice con el 0 o con el +54.
        nombre: { type: string, maxLength: 100, example: Juan }
        apellido: { type: string, maxLength: 100, example: Pérez }
...
    PeticionPerfil:
      type: object
      required: [email, nombre, apellido, telefono]
      description: |
        Los datos personales que alguien puede cambiar de su cuenta. Sin
        contrasena a proposito.
      properties:
        email: { type: string, format: email, maxLength: 254 }
        nombre: { type: string, maxLength: 80 }
        apellido: { type: string, maxLength: 80 }
        telefono: { type: string, example: 11 1234-5678 }
    Usuario:
      type: object
      description: |
        La identidad de login. Nunca incluye el hash de la contraseña ni el
        token de sesión: el token viaja solo en el `Set-Cookie`.
      required: [id, email, telefono, nombre, apellido, creadoEn]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        telefono:
          type: string
          example: "+5491112345678"
          description: |
            En forma canonica E.164. Sale aca porque este esquema responde
            SOBRE VOS —el registro y `/usuarios/yo`—: tu propio telefono lo
            podes ver. El de otra persona sale por otro camino y con otras
            reglas.
        nombre: { type: string }
        apellido: { type: string }
        creadoEn: { type: string, format: date-time }
    UsuarioActual:
      allOf:
        - $ref: '#/components/schemas/Usuario'
        - type: object
          required: [perfilProfesionalId]
          properties:
            perfilProfesionalId:
              type: [string, 'null']
              format: uuid
              description: |
                El perfil profesional del usuario, o `null` si no tiene. No
                tener perfil es normal: es la mitad de los usuarios.
...
  responses:
    PeticionInvalida:
      description: JSON malformado o parámetro con formato inválido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problema' }
...
    NoAutenticado:
      description: |
        No hay sesión, o la que hay venció. Distinto del 403: acá el servidor
        no sabe quién sos.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problema' }
...
    EmailEnUso:
      description: Ya existe una cuenta con ese email
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problema' }
...
    TipoDeContenidoInvalido:
      description: |
        El cuerpo tiene que llegar con `Content-Type: application/json`. La
        regla es parte de la defensa contra CSRF: un formulario de otro sitio
        solo puede mandar `form-urlencoded`, `multipart` o `text/plain`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problema' }
    CuerpoDemasiadoGrande:
      description: El cuerpo de la petición supera el límite permitido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problema' }
    ValidacionFallida:
      description: |
        JSON válido pero datos inválidos. Distinto de 400: "entendí perfecto
        y está mal" es otro problema que "no entiendo lo que mandaste".
      content:
        application/problem+json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Problema'
              - type: object
                required: [errores]